MAINT: Move data.py into helper directory
Moving this module into a helper directory to better describe what it does and that it is shared between modules
