+++ cloud-chatops/src/slack_reminder_api/pr_reminder.py
@@ -2,7 +2,7 @@
 
 from typing import List
 from slack_sdk import WebClient
-from data import PR, Message
+from helper.data import PR, Message
 from read_data import get_config
